Backend.py
O='PhotoshopComfyUI Restarting the server in 5 seconds...'
N='localhost'
T='input'
K=open
I='websocket'
G=True
F=Exception
E=str
B=None
A=print
import asyncio as J,websockets as L,json as H,base64 as P
from PIL import Image as Q
import subprocess as R,os as C,platform as S,os as C,shutil as Y
D={}
class M:
	def __init__(A):
		A.mainDir=C.getcwd()
		for D in range(2):A.mainDir=C.path.dirname(A.mainDir)
		A.tempDir=C.path.join(A.mainDir,'temp');A.inputDir=C.path.join(A.mainDir,T);A.comfyUi=B;A.photoshop=B;A.positive=B;A.negative=B;A.seed=B;A.slider=B;A.image=B;A.mask=B;A.dataDir=B;A.renderDir=B;A.progress=B;A.openWithPS=B;A.QuickEdit=B;A.render_status=B;A.render=B;A.quickSave=B;A.i=0;A.workspace=B
	async def handle_connection(B,websocket,path):
		K='photoshopConnected';J='comfyuiConnected';C=websocket
		try:
			D[C.remote_address]={I:C}
			while G:
				H=await C.recv()
				if H=='imComfyui':
					B.comfyUi=C.remote_address;A('PhotoshopComfyUI comfyUi tab with photoshop node added'+E(B.comfyUi));await B.sendPhotoshop(J,G)
					if B.sendPhotoshop:await B.sendComfyUi(K,G)
				elif H=='imPhotoshop':
					B.photoshop=C.remote_address;A('PhotoshopComfyUI Photoshop launched'+E(B.photoshop));await B.sendComfyUi(K,G)
					if B.comfyUi:await B.sendPhotoshop(J,G)
				elif H=='done':B.sendComfyUi('render_status','genrated')
				elif C.remote_address==B.comfyUi:await B.fromComfyui(H)
				elif C.remote_address==B.photoshop:await B.fromPhotoshop(H)
		except F as L:A(f"PhotoshopComfyUI error handle_connection: {L}");await B.remove_connection(C)
		finally:await C.close()
	async def remove_connection(E,websocket):
		C=websocket
		try:
			del D[C.remote_address]
			if C.remote_address==E.comfyUi:A(f"PhotoshopComfyUI one comfyUi Tab closed {C.remote_address} ");E.comfyUi=B;await C.close()
			elif C.remote_address==E.photoshop:A(f"PhotoshopComfyUI photoshop closed {C.remote_address} ");E.photoshop=B;await C.close()
			else:A(f"PhotoshopComfyUI this ip {C.remote_address} disconnected");await C.close()
		except ValueError:pass
	async def fromPhotoshop(B,message):
		M=message;L='quickSave';J='dataDir';I='workspace'
		try:
			D=H.loads(M)
			if D.get(L):await B.sendComfyUi(L,G)
			if D.get(I):await B.sendComfyUi(I,D.get(I))
			if D.get(J):
				B.dataDir=D.get(J);B.renderDir=C.path.join(B.dataDir,'render.png')
				with K(C.path.join('data.json'),'w')as N:N.write(H.dumps({J:E(B.dataDir)}))
			if not D.get(J)and not D.get(I)and not D.get(L):await B.sendComfyUi('',M)
		except F as O:A(f"PhotoshopComfyUI error fromPhotoshop: {O}");await B.restart_websocket_server()
	async def fromComfyui(B,message):
		X='height';W='width';V='PreviewImage';O='QuickEdit';I=message
		try:
			G=H.loads(I)
			if G.get(V):
				J=G.get(V);D=C.path.join(B.tempDir,J);Z=C.path.join(B.inputDir,J)
				if C.path.exists(D):Y.copyfile(D,Z);await B.sendComfyUi('tempToInput',J)
			elif G.get(O):
				dir=C.path.join(B.mainDir,T,G.get(O).replace('/','\\'))
				if not C.path.exists(dir):A('PhotoshopComfyUI not available',dir)
				else:L,M=Q.open(dir).size;A('PhotoshopComfyUI dir',dir);await B.sendPhotoshop(O,dir);await B.sendPhotoshop(W,L);await B.sendPhotoshop(X,M)
			elif G.get('openWithPS'):
				a=P.b64decode(B.openWithPS);B.i+=1;U='Dolpin_Ai_openWithPS'+E(B.i)+'.psd';A('PhotoshopComfyUI filename',U);D=C.path.join(B.tempDir,U);A('PhotoshopComfyUI file_path',D)
				with K(D,'wb')as N:N.write(a)
				A('PhotoshopComfyUI psd')
				if S.system()=='Darwin':R.call(('open',D))
				elif S.system()=='Windows':C.startfile(D)
				else:R.call(('xdg-open',D))
			else:await B.sendPhotoshop('',I)
		except F as b:A(f"PhotoshopComfyUI error fromComfyui: {b}");await B.restart_websocket_server()
		if I.startswith('rndr'):
			c=P.b64decode(I[4:]);D=f"{B.dataDir}/render.png"
			with K(D,'wb')as N:N.write(c)
			L,M=Q.open(D).size;await B.sendPhotoshop(W,L);await B.sendPhotoshop(X,M)
	async def sendComfyUi(B,name,message):
		C=message
		try:
			if B.comfyUi in D:
				if name=='':await D[B.comfyUi][I].send(E(C))
				else:G=H.dumps({name:E(C)});await D[B.comfyUi][I].send(E(G))
			else:A('PhotoshopComfyUI comfyUi Not Connected')
		except F as J:A(f"PhotoshopComfyUI error sendComfyUi: {J}")
	async def sendPhotoshop(B,name,message):
		C=message
		try:
			if B.photoshop in D:
				if name=='':await D[B.photoshop][I].send(E(C))
				else:G=H.dumps({name:E(C)});await D[B.photoshop][I].send(E(G))
			else:A('PhotoshopComfyUI photoshop Not Connected')
		except F as J:A(f"PhotoshopComfyUI error sendComfyUi: {J}")
	async def restart_websocket_server(C):
		try:
			B=M()
			async with L.serve(B.handle_connection,N,8765):A('PhotoshopComfyUI WebSocket server restarted and waiting for messages')
		except F as D:A(O);J.sleep(5)
async def U():
	try:
		B=M()
		async with L.serve(B.handle_connection,N,8765):await J.Future()
	except F as C:A(O);J.sleep(5)
J.run(U())