requirements.txt
Pillow
asyncio
websockets
